[Integ] build specs to test based components added
For distributions that get published, there might be periods where
plugins have not onboarded to the distribution build for
OpenSearch Dashboards.

We should automatically to this but in the meantime so tests do not
fail because the plugin doesn't actually exist we are opting
for an array and building the specs to run per test run.

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>